feat: dispatch protocol, and the real servers build for wasm32
Two connected steps: the M2 mechanism that the design called the
make-or-break
milestone, and making the real servers and userland compile for the
target it
needs.

**M2 — dispatch protocol (`crates/kernel-wasm`, `crates/wasm-procs`,
`tools/wasm-m2/`)**
- `kernel-wasm` gains the dispatch surface the host loop needs: spawn,
  step,
  syscall, blocked, exit, and `make_endpoint`. It answers questions; it
  does not
  take orders — the process table and run queues stay the kernel's.
- `crates/wasm-procs` is a hand-written guest module, instantiated
  twice, whose
  two entry points block inside `host_syscall` and are resumed by the
  host.
  Straight-line code either side, which is the property §4.2 rests on.
- The host applies the Asyncify protocol measured in the fork spike,
  including
  returning the state to NORMAL when a wait ends so the resumed caller
  does not
  re-enter its rewind path and trap.
- Runs green (7/7). The evidence is the interleaving: A blocks, B runs
  and
  completes the rendezvous, A resumes.

**Real servers and userland build for wasm32 — the toolchain work that
isn't needed**
- §10 claimed this required a new target spec and a std PAL. It does
  not:
  `crates/servers` and `crates/userland` depend on
  `minix-rt`/`minix-std`, all
  `#![no_std]`, and **not** on the fork's `std`, so they build with the
  pinned
  stock toolchain exactly like the kernel. §10 is corrected in place.
- `minix-rt`: the wasm32 syscall gate (`syscall0..6` over a host
  import), and a
  sigreturn trampoline implementing the §6.3 memory-resident contract,
  there
  being no stack for a naked function to read.
- `drivers`: a wasm32 arm for that crate's arch boundary — the same
  single-file
  discipline as the kernel's — pointing at the inert port-I/O and PCI
  surface
  the HAL disposition already prescribes, plus virtio-mmio placeholders.
- `servers`, `userland`: a `uname` arch string, two `asm!("pause")` spin
  hints
  replaced with the portable `core::hint::spin_loop()`, and one more
  `usize::from_le_bytes` over an eight-byte message field.
- `minix-rt`'s `HEAP_LIMIT` needed writing through `u64`: every branch
  of a
  `cfg!` guard is still type-checked, so the 4 GiB literal had to fit a
  32-bit
  `usize` even in the arm wasm32 never evaluates.

**Findings**
- **Endpoints are the kernel's encoding, not handles.** The first
  attempt used
  hand-picked endpoint numbers and deadlocked silently: both sides
  blocked and
  `mini_send` returned OK throughout, because a blocking send and a
  completed
  one are indistinguishable by return value. The host now asks the
  kernel. This
  is the second IPC failure in this port that presents as success.
- The eight-byte-field-read-at-`usize`-width bug has now appeared twice,
  in
  `kernel/src/ipc.rs` and `servers/src/vfs/misc.rs`. Two instances of an
  ABI
  field read at the wrong width is the argument for the per-struct size
  assertion §5.2 proposes, rather than more one-off fixes.

**Docs (`ARCH_WASM32.md`)**
- M2's protocol core recorded as done, §10's toolchain claim corrected,
  the
  32-bit pointer-width invariant added as §5.2 and as risk 7.

Verified: M2 7/7; M0 24/24, M1 4/4, and the fork spike still pass
(re-run
together after the M2 exports landed, and the later edits sit outside
their
dependency graphs). `servers` and `userland` check clean — 0 errors, 0
warnings —
on `wasm32-unknown-unknown` and on all three shipping targets
(`x86_64-pc-minix`, `riscv64gc-unknown-minix`, `aarch64-unknown-minix`),
so every
change here is behaviour-neutral where it matters.

The servers compile for wasm32 but do not yet run: that needs the kernel
to route
the real MINIX syscall numbers and a boot sequence instantiating DS → RS
→ PM →
VFS as modules. No protocol unknowns and no toolchain blockers remain.

+++ crates/servers/src/vfs/misc.rs
@@ -286,7 +286,9 @@ pub fn do_getsysinfo() -> i32 {
     // Read message fields from fs_m_in (mess_lsys_getsysinfo layout)
     let what = i32::from_le_bytes(glob.fs_m_in[0..4].try_into().unwrap_or([0; 4]));
     let dst_addr = u64::from_le_bytes(glob.fs_m_in[4..12].try_into().unwrap_or([0; 8]));
-    let buf_size = usize::from_le_bytes(glob.fs_m_in[12..20].try_into().unwrap_or([0; 8]));
+    // The field holds eight bytes in the message, so read it as u64 and narrow:
+    // `usize::from_le_bytes` would demand four bytes on a 32-bit target.
+    let buf_size = u64::from_le_bytes(glob.fs_m_in[12..20].try_into().unwrap_or([0; 8])) as usize;
 
     // Only superuser may call do_getsysinfo (leaks sensitive data).
     let fp = unsafe { &*glob.fp };
